+++ xlp.ml
@@ -378,7 +378,7 @@ let decl_term_abbrev oc t (k,ntvs,bs) =
   string oc "symbol "; string oc n;
   if ntvs > 0 then
     begin
-      raw_update_tvs_map n ntvs; 
+      raw_update_tvs_map n ntvs;
       string oc " (a0";
       for i=1 to ntvs-1 do string oc " a"; int oc i done;
       string oc " : Set)"
@@ -1116,28 +1116,16 @@ let export_theorem_proof_part b n k =
 (* Functions called in single file generation and by the command "sig"
    in b.mk and Makefile. *)
 
-let types() =
-  let f (n,_) = match n with "bool" | "fun" -> false | _ -> true in
-  List.filter f (types())
-;;
-
 let export_types b =
   export (b^"_types") [] (fun oc -> list decl_typ oc (types()))
 ;;
 
-let constants() =
-  let f (n,_) = match n with "=" | "el" -> false | _ -> true in
-  List.filter f (constants())
-;;
-
 let export_terms b =
-  let n = b^"_terms" in
-  export n [b^"_types"] (fun oc -> list decl_sym oc (constants()))
+  export (b^"_terms") [b^"_types"] (fun oc -> list decl_sym oc (constants()))
 ;;
 
 let export_axioms b =
-  let n = b^"_axioms" in
-  export n [b^"_types"; b^"_terms"]
+  export (b^"_axioms") [b^"_types"; b^"_terms"]
     (fun oc -> decl_axioms oc !the_axioms)
 ;;
 
@@ -1161,10 +1149,10 @@ let export_proofs b r =
 let out_map_thid_name map_thid_name cond with_proof oc =
   MapInt.iter
     (fun k n ->
-      if cond k n
-      then if with_proof then
-             decl_theorem oc k (proof_at k) (DefThmNameAsThmId ("thm_"^n))
-           else decl_theorem oc k (proof_at k) (DeclThmName ("thm_"^n)))
+       if cond k n then
+         decl_theorem oc k (proof_at k)
+           (if with_proof then DefThmNameAsThmId ("thm_"^n)
+            else DeclThmName ("thm_"^n)))
     map_thid_name
 ;;
 
@@ -1175,11 +1163,27 @@ let single_export_theorems f b map_thid_name =
     (out_map_thid_name map_thid_name (fun _ _ -> true) true)
 ;;
 
-(* Called in Makefile by the command "axm". Generate f.lp with, for
+(* Called in Makefile by the command "opam". Generate f.lp with, for
    each named theorem name, a declaration "symbol thm_name : type". *)
 let export_theorems f b map_thid_name cond =
   export f [b^"_types";b^"_terms"]
     (out_map_thid_name map_thid_name cond false)
+;;
+
+(* Called in Makefile by the command "theory". Generate b_theory.lp
+   with, for a type declaration for each type, axiom and named theorem. *)
+let export_theory b map_thid_name =
+  let oc = log_open_out (b^"_theory.lp") in
+  string oc (string_of_file "theory_hol.lp");
+  string oc "\n/* type constructors */\n";
+  list decl_typ oc (types());
+  string oc "\n/* constants */\n";
+  list decl_sym oc (constants());
+  string oc "\n/* axioms */\n";
+  decl_axioms oc !the_axioms;
+  string oc "\n/* theorems */\n";
+  out_map_thid_name map_thid_name (fun _ _ -> true) false oc;
+  close_out oc
 ;;
 
 (* Called in b.mk by the command "part" to create b_part_k and the